reorganized wfg dir
